+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-## [Unreleased]
+## [8.0.25-beta] - 2025-05-08
+- Using Rhino.Inside v8
+- Removed net9 builds
 
-## [1.0.0] - 2024-02-12
+## [8.0.24-beta] - 2025-05-08
+- Strong-named package
+- Added net9 builds
+
+## [8.0.0-beta] - 2024-02-12
 - Initial release